[REF] Change type of the offset to off_t
When accessing storage we need to be able to address even large storages, so
the offset must be of an off_t type.

Fixes:
SEOS-1472

+++ include/OS_FileSystem_int.h
@@ -38,19 +38,19 @@ typedef struct
                         OS_FileSystemFile_Handle_t hFile);
     OS_Error_t (*read) (OS_FileSystem_Handle_t     self,
                         OS_FileSystemFile_Handle_t hFile,
-                        const size_t               offset,
+                        const off_t                offset,
                         const size_t               len,
                         void*                      buffer);
     OS_Error_t (*write)(OS_FileSystem_Handle_t     self,
                         OS_FileSystemFile_Handle_t hFile,
-                        const size_t               offset,
+                        const off_t                offset,
                         const size_t               len,
                         const void*                buffer);
     OS_Error_t (*delete)(OS_FileSystem_Handle_t self,
                          const char*            name);
     OS_Error_t (*getSize)(OS_FileSystem_Handle_t self,
                           const char*            name,
-                          size_t*                sz);
+                          off_t*                 sz);
 } OS_FileSystem_FileOps_t;
 
 /*
